refactor(node): restructure session into session/client and session/server modules
Move session.rs → session/client.rs and session_server.rs →
session/server.rs under a shared session/ module. Re-exports
preserved via pub use session::client::* and session::server::*.
