+++ src/node/session/mod.rs
@@ -1,0 +1,10 @@
+//! Session traits for HOPR node interactions.
+//!
+//! - [`client`] — session client operations (establishing outgoing sessions)
+//! - [`server`] — session server operations (processing incoming sessions)
+
+#[cfg(feature = "node-session-client")]
+pub mod client;
+
+#[cfg(feature = "node-session-server")]
+pub mod server;
